requirements.txt
tinydb
openai>=1.0.0
python-dotenv
requests
beautifulsoup4
